Split JVM audio recorder into Windows and Mac implementations
