+++ composeApp/src/jvmMain/kotlin/de/lehrbaum/voiry/audio/PlatformRecorder.jvm.kt
@@ -1,3 +1,8 @@
 package de.lehrbaum.voiry.audio
 
-actual val platformRecorder: Recorder by lazy { AudioRecorder() }
+import java.util.Locale
+
+actual val platformRecorder: Recorder by lazy {
+	val osName = System.getProperty("os.name")?.lowercase(Locale.getDefault()) ?: ""
+	if ("win" in osName) WindowsAudioRecorder() else MacOsAudioRecorder()
+}
